fix: 계산용 hook 만듦
